check for file existing at URL not server path
The get_sideloaded_file_loc function currently checks the server path for the existence of the file and in a lot of contexts the image will be generated again if it is not found there. Checking for the file URL existing instead will increase compatibility with a lot of CDN situations including WPEngine's LargeFS system, and I assume would also make things work better with the [WP Offload Media](https://deliciousbrains.com/wp-offload-media/) plugin.